Reorganize datalad paths to be more consistent post DataLad backend.
