+++ packages/openneuro-app/src/scripts/datalad/dataset/dataset-main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import ClickToEdit from '../common/forms/click-to-edit.jsx'
+import ClickToEdit from '../../common/forms/click-to-edit.jsx'
 import DatasetRoutes from './dataset-routes.jsx'
 
 const DatasetMain = ({ dataset }) => (
