+++ packages/openneuro-app/src/scripts/datalad/dataset/dataset-page.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import LeftSidebar from './left-sidebar.jsx'
 import LeftSidebarButton from './left-sidebar-button.jsx'
 import DatasetMain from './dataset-main.jsx'
-import Tools from './fragments/tools/dataset-toolbar.jsx'
+import Tools from '../fragments/tools/dataset-toolbar.jsx'
 
 export default class DatasetPage extends React.Component {
   constructor(props) {
@@ -33,7 +33,7 @@ export default class DatasetPage extends React.Component {
             sidebar={this.state.sidebar}
             toggle={this.toggleSidebar}
           />
-          <Tools dataset={this.props.dataset}/>
+          <Tools dataset={this.props.dataset} />
           <div className="fade-in inner-route dataset-route light">
             <div className="clearfix dataset-wrap">
               <div className="dataset-animation">
