+++ packages/openneuro-app/src/scripts/datalad/dataset/left-sidebar.jsx
@@ -25,7 +25,7 @@ SidebarRow.propTypes = {
   datasetId: PropTypes.string,
   id: PropTypes.string,
   version: PropTypes.string,
-  draft: PropTypes.boolean,
+  draft: PropTypes.bool,
 }
 
 const LeftSidebar = ({ datasetId, snapshots }) => (
